feat(inspector): promote from experimental to regular, warn users to update config
