+++ .changeset/bright-fans-count.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/vite-plugin-svelte': minor
+---
+
+feat(inspector): Promote experimental.inspector to regular option
